Move access token related implementation classes to separate package
