+++ graylog2-server/src/main/java/org/graylog2/bindings/PersistenceServicesBindings.java
@@ -35,11 +35,11 @@ import org.graylog2.notifications.NotificationServiceImpl;
 import org.graylog2.savedsearches.SavedSearchService;
 import org.graylog2.savedsearches.SavedSearchServiceImpl;
 import org.graylog2.security.AccessTokenService;
-import org.graylog2.security.AccessTokenServiceImpl;
 import org.graylog2.security.MongoDBSessionService;
 import org.graylog2.security.MongoDBSessionServiceImpl;
 import org.graylog2.security.ldap.LdapSettingsService;
 import org.graylog2.security.ldap.LdapSettingsServiceImpl;
+import org.graylog2.security.token.AccessTokenServiceImpl;
 import org.graylog2.shared.users.UserService;
 import org.graylog2.streams.StreamRuleService;
 import org.graylog2.streams.StreamRuleServiceImpl;
